Set secondary boot mode register
On the Zynqmp, boot can occur on a secondary device different from the primary
boot device set by the boot pins.  This is configured via the boot_device
parameter in bootgen's BIF file.

When used, the FSBL still does not set the BOOT_MODE_USER register, and
therefore secondary SBLs (U-Boot) incorrectly report the boot device as the
original.  This patch fixes that.

In addition, the FSBL does not correctly compare the boot device reported in
the bootgen image header to the boot device reported by the BOOT_MODE_USER
register.  Both give an ordinal that corresponds to the boot device, but the
ordinals don't match up and must be translated.

Compare:
    Zynq UltraScale+ Device TRM UG1085 (v2.1) August 21, 2019,
        Table 11-1: Boot Modes
with:
    Bootgen User Guide UG1283 (v2019.2) October 30, 2019,
        Table 13: Zynq UltraScale+ MPSoC Device Image Header Table
            Secondary Boot Device field

Note: the latter table also has a mistake - all ordinals from "4 - SD1" and
thereafter should be incremented by one.