switch dependency of u-boot-uenv to iveia-boot
Planned future changes will allow a variable like IV_IO=blackwing to create a
DTB file that includes both MACHINE and IO board specs.  To keep the iveia-boot
recipe independent from the iveia-image-* recipes, we need to have uEnv.txt be
a dependency of iveia-boot.  This will allow bitbakes of the form:
    MACHINE=atlas-ii-z8 IV_IO=blackwing bitbake iveia-boot

This change just fixes the dependency, and simplifies uEnv.txt creation.