u-boot: fixed mac addr failure when already set
If savenv is used, then the mac address vars (eth#addr) get saved.  On reboot
when trying to set the mac addresses from the SNs, error messages are shown
because the eth#addr vars are not allowed to be rewritten.

With this fix, the eth#addr vars are not set by SN if they were already set.

Also, the more canonical eth_env_set_enetaddr() is used.