version: added Linux version banner
iveia-version-header.bbclass did not work for Linux, because when Linux patches
the source, it commits the changes (whereas other recipes patch the source but
leave the work dir dirty).  This causes the "git describe"d version to report
a false commit (of the last patch).

The fix is to get the version just before patch time.