dt overlays: create dt overlay and combo dtb
Now, 3 DTs are created: a mainboard DTB, a ioboard DTBO, and a combo DTB that
includes both.  The mainboard/ioboard DTBs are used for Linux, loaded at U-Boot
time.  The combo is used to be burned into U-Boot via EXT_DTB.  This allows
U-Boot to have the ioboard DT info, even though U-Boot doesn't support building
in a DT overlay.

To create the combo DTS, the env var IVIO must be set to the ioboard (or set as
in a conf file)