openamp: fix broken dependency
Openamp requires xsct but does not DEPEND on it.  Specifically, recipes in
open-amp, such as:

meta-xilinx-tools/recipes-openamp/examples/openamp-fw-echo-testd_git.bb

require xsct.  But, if being built alone, they fail to download xsct, because
of the dependency handling in xsct-tarball.bbclass.  The above recipe inherits
xsctapp, which thorugh a chain ends up inheriting xst-carball.bbclass, which
has a function xsct_event_extract().  In that function, xsct is only downloaded
if it matches one of XSCT_TARGETS, which it doesn't.

Note that normally this isn't an issue, because I suspect most users would
bitbake petalinux-image-full, which itself depends on other XSCT_TARGETS.  But
we have our own layer that builds the rootfs WITHOUT dependencies on FSBL and
other targets in XSCT_TARGETS.  Therefore xsct is not downloaded and the above
recipe fails.

For example, to reproduce this in 2019.2 tools, from a clean start (i.e. rm -rf
build/tmp) and run:

bitbake openamp-fw-echo-testd