device-tree: fix broken overlay loading
The main dtb file must be compiled with the "-@" option, in order to include
the symbols section.  If not, then the dtbo overlays will not apply in U-Boot
("fdt apply"), with the failure:
    failed on fdt_overlay_apply(): FDT_ERR_NOTFOUND
    base fdt does did not have a /__symbols__ node
    make sure you've compiled with -@

Strangely, previously built and released dtb files were built with symbols,
which can be verified by decompiling the dtbs.  So, it's unclear how/where this
issue started appearing, as reverting to older commits and rebuilding showed
this same issue.

Finally, the DTC flags used were pulled from the core devicetree.bbclass.