mdio-i2c-gen: add driver for I2C based MDIO support on Blackwing carrier
This adds a generic driver for I2C-based MDIO support.

Note, this driver can not be mainlined as-is to Linus' Linux, as it requires a
hack: the I2C driver probe must run before this driver's probe.  There is no
current framework to do this. See mdio-i2c-gen-change-*.patch for details.

This commit adds DTS support for the iVeia Blackwing carrier.