new ivinstall recipe
ivinstall is a new recipe that merges Rick's self-extracting archive with the
devkit's sdformat to provide all-in-one installation utility.

ivinstall wraps all of the final images required to boot and run Linux into a
single self-extracting shell archive.  The archive can be run from a host
(Linux or mac) or the target.  It supports 3 primary modes for installing
images:
    - directly onto an SD card
    - over JTAG-only to the target, and onto the target's SD or eMMC
    - remote to the target over ssh, and onto the target's SD or eMMC
It supports sdformat's original options for formatting, copying, setting
partition sizes, etc.

QSPI reflash is not currently implemented, but included in the options.

Not all options (e.g. formatting) are supported on mac.