atlas-ii-z8ev: memory timing
atlas-ii-z8ev DDR can be one of:
    1) MT40A512M16JY-083E IT:B (DDR-2400 16-16-16)
    2) MT40A512M16LY-062E IT:E (DDR-3200 22-22-22)

In both cases, we run at 1200MHz, i.e. DDR-2400.  For the slower part (#1),
from the datasheet:

    tAA = tRCD = tRP = 13.32ns --> 1200MHz * 13.32ns = 15.94 clocks (i.e. 16)

For the faster part (#2), which is running at same 1200MHz (DDR-2400) instead
of 1600MHz (DDR-3200):

    tAA = tRCD = tRP = 13.75ns --> 1200MHz * 13.75ns = 16.50 clocks (i.e. 17)

So the faster part needs to use a 17-17-17 timing setup (which is also
compatible with the slower part).

P.S. This patch fixes the issue where we have seen intermittent boot issues
with some boards, where the FSBL would hang during memory initialization.