QSPI32: Workaround for unfinished QSPI transfer
In some cases, XFsbl_Qspi32Copy() returns before the transfer is complete.

Reason unknown, but specific case was when cold booting via QSPI24, and then
using the CRL_APB_BOOT_MODE_USER register to soft reboot in QSPI32 mode.  The
XFsbl_Qspi32Copy() was called multiple times to read the bootgen image header
and partitions, and one of the copies would fail (always the same one).

This is a workaround, not a clean fix - but even a negligible usleep(0)
resolves the issue.  We sleep a bit longer to be safe.  This adds a negligible
time to the total boot time.