fsbl: added QSPIxx boot modes
Added QSPI24 (3-byte addressing) and QSPI32 (4-byte addressing) boot modes.
QSPI24 was already implicitly used, as it is the bootstrap mode.  But QSPI32
was added so that it is now possible to change the boot sequence to:

iv_boot_sequence_t boot_seq[] = {
    { .mode = XFSBL_QSPI32_BOOT_MODE, .num_attempts = 1 },
};

...which causes the primary boot mode to be QSPI24, and then a soft reset into
QSPI32 mode.  This might be useful for the case where the device boots solely
from QSPI, and software changes the mode to 4-byte addressing.  In this case,
it has been reported (by customers and in Xilinx forums) that a soft reset back
into QSPI24 mode will not work unless the QSPI itself is reset.

This commit also resolves some broken register writes which were previously
only working by quirk of fate.