ivinstall: fix fail waiting for partitions to appear
After partitions are created via parted, they may not appear immediately.  So
we wait, and verify the appear via lsblk.  Even when using this method, it is
possible on some systems for partitions to appear via a call to lsblk, and
dissapear on the call immediately after.

Therefore, we reconfirm the partitions after a sleep.

Additionally, we supress benign error messages that can occur from lsblk.