QSPI32 is now default fallback boot mode
iVeia boards boot from QSPI flash, and are configured in QSPI24 via the boot
mode pins.  However, the standard flash size on boards is 256MB.  Access to
memory above 16MB conflicts with the FSBL QSPI flash driver code (it partially
works, but transfers across the 16MB bounday fail).

To resolve this, the default fallback for QSPI is QSPI32.  So, when booting
fully from QSPI, the device will start in QSPI24 mode and then soft reset into
QSPI32 mode.  At that point, the full memory is accessible normally with the
FSBL driver.

This has been tested with a 17MB bitfile inserted as the third boot.bin entry
(after FSBL and PMUFW).  Without this change, the boot hangs during the bitfile
load.

The 17MB bitfile was created with slew of adders and multipliers, and with
compression turned off.