ivinstall: changed JTAG header from 64 to 32 bits
To be compatible with both zynq and zynqmp, ivinstall now uses 32-bit based
JTAG header values.

ivinstall uses some hardcoded trickery to load over JTAG: it attaches a 3 value
header to the images loaded to be read at U-Boot or Linux time.  Because U-Boot
needed to read them and process them with 'setexpr', they needed to be 64-bit
because setexpr is broken for 32-bit values.

This commit adds a pacth to fix setexpr, and then adjusts the addresses to use
32-bit based offsets.

Tested with zynq and zynqmp using -Q option, but not -J option.