atlas-i-z7e: enabled usb
USB 0/1 enabled in XSA and DTS - note pinctrl node is required.

Tested on atlas-i-z7e on Captiva, on USB0 (micro-AB) with USB-ID (pin 4)
shorted to ground.  The short was required because I don't have a USB micro-A
cable, which has the short internally.  And the micro-A cables are tough to
find!

USB1 was not working, at least on the engineering board I had (A2VKR).