helios-z7-pcie: Booting to Linux via ivinstall -joJ
Two issues were solved:

On Windows, if the ivinstall.tcl script is run via ssh and
immediately exits, the JTAG connection becomes hung (process is
halted).  As a workaround, we test for Linux/Windows, and run a short
sleep after running the script.  The sleep must be run AFTER xsdb
exits, but before ssh exits.

The ivinstall image is stored above the memory used by Linux (via mem=xxx).
But unfortunately on arm, U-Boot by default assumes that it can relocate the
initrd/fdt to high memory.  This conflicts with Linux, which expects the
initrd/fdt addresses to be within it's addressable memory.  Therefore, the
initrd_high/fdt_high parameters are used to relocate those images to an
allowable area.  This must be within the Linux memory set with kernel_mem_mb.