ivinstall: In JTAG mode leave more Linux mem
In JTAG mode, ivinstall puts the ivinstall script above Linux memory.
Previously, the split was 256MB for Linux, and the rest for the ivinstall
script.  The assumption is there is at least 512MB on the device.

256MB is not enough memory to boot, copy the ivinstall script to tmpfs, and
then unpack the ivinstall script.  So, the new split is at 384MB.  For current
ivinstall images (~50MB), this works fine, and should hopefull give enough
space to roughly double the image size.