u-boot-uenv: Added pauseboot command.
The command 'run pauseboot' will run the boot seqeunce up to the first uEnv.txt
that it finds, and execture everything in the uEnv.txt up to the final
'boot_kernel' command.

Useful to load all images and settings, and then modify or review the settings
before running the final boot command.