u-boot-uenv: Zynq fix for setting SD boot device.
Zynqmp board support automatically sets a variable 'sdbootdev' that is the
device number (0 or 1) of the SD card the target is booting from.  This is
passed to the kernel to select the root device.  On Zynq, 'sdbootdev' is not
set.

However, in both cases, we know which device we're booting from as we try them
in sequence, according to the 'boot_targets'.  So, we now just manual set the
'sdbootdev' whether on zynq or zynqmp.