Fix common JTAG load failures after powercycle
WORKAROUND: Often, when trying to install on a target that has just
been powercycled, JTAG loads can fail.  The issue appears related to
letting the target run fully into Linux (but this is not always
true).  Because the extract_archive_to_TMPDIR function takes a while
to run, the actual load doesn't happen until far along into the boot.
The fix is to run a script right NOW that halts the processor, before
Linux has a chance to boot.