download script: allow DONTMOVELAYER argument
By default, the download script moves the meta-iveia repo into a subdirectory
where all of the other layers reside (Xilinx's convention).  If meta-iveia is
included as a submodule of another repo, this breaks meta-iveia, because you
can't move submodules (without some work - submodules have a text link back to
the parent repo as their .git).

The download script now accepts a DONTMOVELAYER to keep the layer where it is.