allow FSBL boot by SoM PN
The boot sequencer now allows boot by part number, by setting num_attempts = -1.

If set, then Multiboot will be used with the iVeia SoM part number ordinal.
For example, if the board was an iVeia atlas-ii-z8ev (which has a part number
ordinal of 00127), and if this was booting via SD, the sequencer would attempt
to load the file BOOT0127.BIN.

This is currently configured for external SD (SD0) only.  The current sequence
is:
- Try BOOT.BIN on SD0 (external SD), one attempt.
- Try BOOTNNNN.BIN on SD0 (external SD), where NNNN is the SoM PN
  (205-NNNNN-VV-RR).  Note: 5 digit part number is truncated to 4 digits.
- Try BOOT.BIN on SD1 (external SD), one attempt.
- Try QSPI32, one attempt - should always work.