fsbl: move iv_z8_init() hook to FSBL hook file
Previous method for including iv_z8_init() hook was to patch psu_init.c.
However, this is fragile, as psu_init.c can change based on the components in
the XSA, causing the patch to fail.

The better approach is to include the hook in the XFsbl_PsuInitHook(), as
intended.  A minor risk is that now iv_z8_init() occurs at a slightly later
time (after psu_init(), instead of NEAR the end of it), which has a slight
possibility of effecting existing usages.