00130: Use alternate EEPROM IPMI format
The 00130, in some cases, uses a non-iVeia EEPROM IPMI format.  As such, the
board SN is not parseable by standard code, and an iVeia MAC address cannot be
used.

A special case was added, that is only compiled in for the 00130 board, and is
only used if the iVeia IPMI format is not found.

To do this, a Kconfig was added for iVeia boards.  It can be used for future
CONFIG_* vars as well.