ivinstall: Split -J JTAG_REMOTE into two options
JTAG_REMOTE previously took a double argument (separated by a colon): the
remote machine, and the adapter used on that machine.  However, this interface
doesn't work if the user is on a local JTAG machine and still needs to specify
the adapter.  So now it is split into -J JTAG_REMOTE and -a JTAG_ADAPTER
options.