ivinstall: fix failed mkfs after parted
Parted has a known issue that it may remount a filesystem after creating the
partition if the old filesystem data is still there.  Only fix is to wipe the
full block device (way slow) or unmount again - so we unmount again.

Note: we had not previously seen this occur until the upgrade to Xilinx 2021
tools.

Discussed here, and other places:

https://unix.stackexchange.com/questions/432850/is-parted-auto-mounting-new-partitions