ivinstall: Modify -b SIZE option defaults
The -b SIZE option used to have a default that varied depending on the SD card
size.  However, in order to create a future JTAG (-j) format option where the
SD card is used for temporary storage, we'll need to know the P2_START position
from the host - but the host doesn't have access to the target's SD card.

To fix this, we now always use the same 4GB default first partiton size, and
users can manually change it if the card is smaller, which is rare nowadays
anyway.