ivinstall: new JTAG install method for large images
ivinstall for JTAG was broken for larger images.  This was exacerbated by the
change to 2021 tools, where the images increased significantly in size.

The new method is described in the ivinstall-header.sh.  In brief:
    - Images are copied over JTAG into RAM
    - large ivinstall image is copied onto SD (does not stay in RAM as
      previously)
    - Linux boots
    - ivinstall is moved to a new FAT partiton on SD
    - ivinstall formats card

This method should work for ivinstall images up to 1GB or larger, depending on
RAM and SD size.  Current images are around 100MB.

Drawback is that JTAG installs are now required to reformat the card (i.e -f
must be specified), as the process uses the SD card temporarily and therefore
overwrites images on the card.