z7x: configure Eth MDIO
On z7x, Eth MDIO is controller via GPIO on an I2C expander.  This patch adds
the MDIO_GPIO driver and device tree config.

Note that the MDIO I2C GPIOs are on the SoM, and are therefore defined in the
SoM's device tree, but the actual Eth PHY is on the carrier and therefore
defined in the overlay.