z7: add fpga-manager to dynmically load bitfiles
For Zynq devices, fpgautil can be used to load FPGA images.  However, unlike on
ZynqMP, it does not support bitfiles, only binfiles.  The files differ only by
header/trailer bytes.

To convert a bitfile to a binfile run:
    # echo "all:{mybitfile.bit}" > temp.bif
    # bootgen -image temp.bif -arch zynq -process_bitstream bin

which will create mybitfile.bit.bin, which can then be loaded on the target
with:
    # fpgautil -b mybitfile.bit.bin

Also note that in Yocto, bootgen is available at:
    tmp/sysroots-components/x86_64/bootgen-native/usr/bin/bootgen