ivinstall: create IV_TAR_FILE_MAP as a static variable
Previously, the list of tarfiles to include in the final image was created
dynamically in do_deploy.  This was not necessary, as it was a static list, and
could be created as an enviroment variable (with overrides).

The issue was that for someone to add additional files to the image in their
own layer, they'd need to override the whole do_deploy.  Now, you can just
append the IV_TAR_FILE_MAP_append variable, for example just add a
ivinstall-minimal.bbappend containing:

IV_TAR_FILE_MAP_append = "\
    ${DEPLOY_DIR_IMAGE}/xilinx.bit         :   boot/xilinx.bit \
    "

And the above xilinx.bit will be added to the ivinstall image.