zap: removed extraneous call to arch_setup_dma_ops()
This call had no effect, but was incorrect - likely added because of an
incorrect device tree.  It has been in place since 2017.  The removal is
required for this code to build on z7 boards (z8 boards had a hacked in patch
to get this to work - the patch is now removed).

Portable drivers should not directly be be making arch_* calls anyway - in this
case, the arch_setup_dma_ops() is indirectly called automatically through
of_dma_configure_id() given a good device tree (the device tree DMA entries are
created automatically via Xilinx's DTG).