initrd: rearrange boot mem layout
When ZAP driver was permanently added, ZAP memory (which is reserved in DTB)
was blocking normal initrd load in U-Boot, even though U-Boot doesn't use ZAP.
Rerranged layout to make room for both ZAP and initrd.

All of the boot files and ZAP fit in the first 512MB, which is currently plenty
of space.  If there is not enough room in the future, ZAP could be moved
elsewhere (above 512MB).

This also fixes "ivinstall -j", which required initrd.