DT: split z7/z8 device tree overlays
Previously, a simple file suffix ("-z7") was used to separate z7 and z8
overlays.  They are separate because they are not compatible.

This clutters the IO namespace.  In addition, this creates confusion when
installing via ivinstall - if you install on a z7 SoM and use an overlay
without the "-z7" extension, the install works, but the board generally won't
boot.

Now, z7/z8 overlays are kept separate, and only compatible overlays are
packaged within the ivinstall image.