xsa: update instructions for modifying XSA after 2019.2
In the past, the Xilinx XSA file contained a Block Design TCL script that could
be used to recreate the Block Design from which the XSA was created.  After
Vivado 2019.2, this feature was removed - the TCL script is gone, and there is
no way to recreate a Vviado XPR project from an XSA.

Unfortunately this means that a Block Design TCL script must be saved alongside
- and kept in sync with - the system.xsa.

This update changes the README instructions to reflect this.