fix broken SRC_URIs
If a git branch name is not specified in a SRC_URI, bitbake (in 2021.2)
defaults to "master".  The libvpx and mobile-broadband-provider-info repos
renamed their "master" branch to "main".  This leads to a fetch error if the
repo is not cached in DL_DIR.  Append these recipes to explicitly specify
'branch=main'.

These appends can be removed when the source layers are updated.